+++ .env.example
@@ -5,7 +5,7 @@ BOT_ADMINS=123,456
 AUTOROLE=MSG_ID:ROLE_ID:EMOJI:AUTOREMOVE
 # Another example: AUTOROLE=738932146978160661:728202487672078368:❌:false,738932146978160661:738936540465725520:✅:true
 
-DATABASE_URL="localhost:5432/tsc-bot"
+DATABASE_URL="postgres://tscbot:tscbot@localhost:5432/tscbot"
 
 # Role given to trusted members, not full moderators, but can use some commands which
 # are not given to all server members.
@@ -15,10 +15,11 @@ RULES_CHANNEL=
 
 ROLES_CHANNEL=
 
-HELP_CATEGORY=
 HOW_TO_GET_HELP_CHANNEL=
 HOW_TO_GIVE_HELP_CHANNEL=
-GENERAL_HELP_CHANNEL=
+
+HELP_FORUM_CHANNEL=
+HELP_REQUESTS_CHANNEL=
 
 # Time in milliseconds before !helper can be run
 TIME_BEFORE_HELPER_PING=
