+++ src/env.ts
@@ -16,10 +16,10 @@ export const autorole = process.env.AUTOROLE!.split(',').map(x => {
 
 export const dbUrl = process.env.DATABASE_URL!;
 
-export const helpCategory = process.env.HELP_CATEGORY!;
 export const howToGetHelpChannel = process.env.HOW_TO_GET_HELP_CHANNEL!;
 export const howToGiveHelpChannel = process.env.HOW_TO_GIVE_HELP_CHANNEL!;
-export const generalHelpChannel = process.env.GENERAL_HELP_CHANNEL!;
+export const helpForumChannel = process.env.HELP_FORUM_CHANNEL!;
+export const helpRequestsChannel = process.env.HELP_REQUESTS_CHANNEL!;
 
 export const trustedRoleId = process.env.TRUSTED_ROLE_ID!;
